피드백 수정완료

+++ src/main/java/com/sparta/schedule/service/ScheduleService.java
@@ -18,51 +18,51 @@ public class ScheduleService {
         this.scheduleRepository = scheduleRepository;
     }
 
-
     public ScheduleResponseDto createSchedule(ScheduleRequestDto requestDto) {
-        // RequestDto -> Entity
         Schedule schedule = new Schedule(requestDto);
-
-        // DB 저장
         Schedule saveSchedule = scheduleRepository.save(schedule);
-
-        // Entity -> ResponseDto
         ScheduleResponseDto responseDto = new ScheduleResponseDto(schedule);
-
         return responseDto;
     }
 
-    // 전체조회
-    public List<ScheduleResponseDto> getSchedules() {
-        // DB 조회 (내림차순)
-        return scheduleRepository.findAllByOrderByModifiedAtDesc().stream().map(ScheduleResponseDto::new).toList();
+    public List<Schedule> getSchedules() {
+        return scheduleRepository.findAllByOrderByModifiedAtDesc();
     }
 
-    // 선택한 일정조회
-    public List<ScheduleResponseDto> getSchedule(Long id) {
-        // 해당 일정이 DB에 존재하는지 확인
+    public ScheduleResponseDto getSchedule(Long id) {
         Schedule schedule = findSchedule(id);
-        // DB 조회 선택한 일정 조회
-        return scheduleRepository.findById(id).stream().map(ScheduleResponseDto::new).toList();
+        ScheduleResponseDto responseDto = new ScheduleResponseDto(schedule);
+        return responseDto;
     }
 
     @Transactional
-    public ScheduleResponseDto updateSchedule(Long id,ScheduleRequestDto requestDto,Long password) {
+    public ScheduleResponseDto updateSchedule(Long id,ScheduleRequestDto requestDto) {
         // 해당 일정이 DB에 존재하는지 확인
         Schedule schedule = findSchedule(id);
         // 일정 수정
+        if(checkPWAndFindSchedule(schedule, requestDto)) {
         schedule.update(requestDto);
-        scheduleRepository.save(schedule);
+        } else {
+            throw new IllegalArgumentException("패스워드가 일치하지 않습니다.");
+        }
         ScheduleResponseDto scheduleResponseDto = new ScheduleResponseDto(schedule);
         return scheduleResponseDto;
     }
 
-    public Long deleteSchedule(Long id) {
+    public void deleteSchedule(Long id,ScheduleRequestDto requestDto) {
         // 해당 일정이 DB에 존재하는지 확인
         Schedule schedule = findSchedule(id);
-        // 일정 삭제
-        scheduleRepository.delete(schedule);
-        return id;
+        if(checkPWAndFindSchedule(schedule, requestDto)) {
+            scheduleRepository.delete(schedule);
+        }
+    }
+
+    private boolean checkPWAndFindSchedule (Schedule schedule, ScheduleRequestDto requestDto) {
+        if(schedule.getPassword().equals(requestDto.getPassword())){
+            return true;
+        } else {
+            throw new IllegalArgumentException("비밀번호가 일치하지 않습니다.");
+        }
     }
 
     private Schedule findSchedule(Long id) {
